# Add check-network-performance.sh