Fix #665 partially. 

**사용 방법**
son@son:~/go/src/github.com/cloud-barista/cb-tumblebug/src/testclient/scripts/sequentialFullTest$ 

1. `./create-all.sh` all 1 shson02 ../testSetCombination.env 로 MCIS 생성.


2. `./check-network-performance.sh` all 1 shson02 ../testSetCombination.env

**출력 결과 예시 (MCIS 내부의 VM들을 PrivateIP 로 Ping 테스트)**

```
####################################################################
## Check network performances in MC-Infra (fping, iperf)
####################################################################
[Check jq and putty-tools package (if not, install)]

[GENERATED PRIVATE KEY (PEM, PPK)]

[MCIS INFO: mc-shson02]
IPLIST: 54.255.143.90 52.221.199.111 18.136.103.253 104.155.192.147 35.201.238.215 34.80.168.89
PRIVIPLIST: 192.168.2.200 192.168.2.204 192.168.2.107 192.168.3.4 192.168.3.2 192.168.3.3
IPCOMBLIST: 54.255.143.90 192.168.2.200 52.221.199.111 192.168.2.204 18.136.103.253 192.168.2.107 104.155.192.147 192.168.3.4 35.201.238.215 192.168.3.2 34.80.168.89 192.168.3.3
....

**[Ping to all nodes in MC-Infra (CMD: fping -e 192.168.2.200 192.168.2.204 192.168.2.107 192.168.3.4 192.168.3.2 192.168.3.3)]**

[aws-ap-se-1-0 (54.255.143.90)] ping to 
192.168.2.107 is alive (0.22 ms)
192.168.2.200 is alive (0.04 ms)
192.168.2.204 is alive (0.16 ms)
192.168.3.2 is alive (49.0 ms)
192.168.3.3 is alive (49.2 ms)
192.168.3.4 is alive (43.5 ms) 

[aws-ap-se-1-1 (52.221.199.111)] ping to 
192.168.2.107 is alive (0.23 ms)
192.168.2.200 is alive (0.14 ms)
192.168.2.204 is alive (0.03 ms)
192.168.3.2 is alive (48.4 ms)
192.168.3.3 is alive (48.6 ms)
192.168.3.4 is alive (43.2 ms) 

[aws-ap-se-1-2 (18.136.103.253)] ping to 
192.168.2.107 is alive (0.03 ms)
192.168.2.200 is alive (0.23 ms)
192.168.2.204 is alive (0.22 ms)
192.168.3.2 is alive (48.0 ms)
192.168.3.3 is alive (48.7 ms)
192.168.3.4 is alive (43.4 ms) 

[gcp-as-east1-0 (104.155.192.147)] ping to 
192.168.2.107 is alive (42.2 ms)
192.168.2.200 is alive (42.2 ms)
192.168.2.204 is alive (42.2 ms)
192.168.3.2 is alive (1.57 ms)
192.168.3.3 is alive (1.98 ms)
192.168.3.4 is alive (0.04 ms) 

[gcp-as-east1-1 (35.201.238.215)] ping to 
192.168.2.107 is alive (47.6 ms)
192.168.2.200 is alive (48.5 ms)
192.168.2.204 is alive (47.5 ms)
192.168.3.2 is alive (0.07 ms)
192.168.3.3 is alive (1.86 ms)
192.168.3.4 is alive (0.28 ms) 

[gcp-as-east1-2 (34.80.168.89)] ping to 
192.168.2.107 is alive (47.6 ms)
192.168.2.200 is alive (48.4 ms)
192.168.2.204 is alive (47.7 ms)
192.168.3.2 is alive (0.31 ms)
192.168.3.3 is alive (0.07 ms)
192.168.3.4 is alive (0.37 ms)

Done!
```